% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calc_counts}
\alias{calc_counts}
\title{calc_counts
generate counts of a column}
\usage{
calc_counts(
  dt,
  target,
  target_name = target,
  treat,
  indent = nbsp(n = 4L),
  .total_dt = NULL,
  pct_dec = 1
)
}
\arguments{
\item{dt}{table input to use.}

\item{target}{string, column name to calculate count statistics on}

\item{target_name}{string, how to display column name when new table is create, if null then target column name is used}

\item{treat}{string, treatment population to use as row headers}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}

\item{.total_dt}{optional table for total counts to be derived}

\item{pct_dec}{decimal places for percentages}
}
\value{
a list containing a data.table
}
\description{
calc_counts
generate counts of a column
}
\examples{
adsl <- random.cdisc.data::cadsl
RACE<-calc_counts(dt = adsl, 'RACE', target_name = 'RACE', treat = 'ARM', indent = '')
}
