% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calc_desc}
\alias{calc_desc}
\title{calc_desc}
\usage{
calc_desc(
  dt,
  target,
  target_name = target,
  treat,
  indent = nbsp(n = 4L),
  pct_dec = 1
)
}
\arguments{
\item{dt}{table input to use.}

\item{target}{string, column name to calculate descriptive statistics about}

\item{target_name}{string, how to display column name when new table is create, if null then target column name is used}

\item{treat}{string, treatment population to use as row headers}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}
}
\value{
a list containing a data.table
}
\description{
Given a table, calculate descriptive statistics for a target column.
}
\examples{
adsl<- random.cdisc.data::cadsl
age<-calc_desc(adsl, 'AGE', 'AGE', treat='ARM')

}
