% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calc_stats}
\alias{calc_stats}
\alias{calc_stats.numeric}
\alias{calc_stats.character}
\alias{calc_stats.factor}
\alias{calc_stats.logical}
\title{calc_stats}
\usage{
calc_stats(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)

\method{calc_stats}{numeric}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)

\method{calc_stats}{character}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)

\method{calc_stats}{factor}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)

\method{calc_stats}{logical}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)
}
\arguments{
\item{dt}{table input to use.}

\item{target}{string, column name to calculate count statistics on}

\item{target_name}{string, how to display column name when new table is create, if null then target column name is used}

\item{treat}{string, treatment population to use as row headers}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}
}
\value{
a list containing a data.table
}
\description{
calc_stats
}
\examples{
x
}
