% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{event_count_by}
\alias{event_count_by}
\title{count total, at least one and splits on the treatment, grouping and target variables}
\usage{
event_count_by(
  dt,
  patient,
  treat,
  rows_by,
  target,
  .total_dt = NULL,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)
}
\arguments{
\item{dt}{table to perform function on}

\item{patient}{string, column for unique patient identifier}

\item{treat}{string, treatment population to use as column headers}

\item{rows_by}{string, grouping variable to split events by.}

\item{target}{string, target column to provide further counts within group provided by rows_by}

\item{.total_dt}{optional table for total counts to be derived}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}
}
\value{
data.table
}
\description{
count total, at least one and splits on the treatment, grouping and target variables
}
\examples{
adae <- random.cdisc.data::cadae |> dplyr::filter(ANL01FL == "Y")
event_count_by(dt=adae, patient='USUBJID', treat='ARM', target='AEDECOD',rows_by = 'AEBODSYS',.total_dt = adsl, indent = ' ')

}
