% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{merge_table_lists}
\alias{merge_table_lists}
\title{merge_table_lists}
\usage{
merge_table_lists(dt_l)
}
\arguments{
\item{dt_l}{nested list of data.table outputs from created from other dtlg functions}
}
\value{
data.table
}
\description{
merge_table_lists
}
\examples{
x
}
