% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{multi_event_true}
\alias{multi_event_true}
\title{multi_event_true}
\usage{
multi_event_true(
  dt,
  event_vars,
  patient,
  treat,
  label = NULL,
  .total_dt = NULL,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)
}
\arguments{
\item{dt}{table to perform function on}

\item{event_vars}{events to perform counts on}

\item{patient}{string, column for unique patient identifier}

\item{treat}{string, treatment population to use as row headers}

\item{label}{row name to be displayed, if NULL then column name or label is used}

\item{.total_dt}{optional table for total counts to be derived}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}
}
\value{
list of data.table containing events
}
\description{
multi_event_true
}
\examples{
x
}
