% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTables.R
\name{summary_table}
\alias{summary_table}
\title{summary_table}
\usage{
summary_table(
  dt,
  target,
  treat,
  target_name = NULL,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)
}
\arguments{
\item{dt}{table to perform function on}

\item{target}{vector of column names desired to obtain information on}

\item{treat}{string of treatment variable used for splitting / grouping data}

\item{target_name}{names to be displayed for target columns either NULL, which uses target, of vector of desired names for all target columns}

\item{indent}{indent to be used for display and formatting purposes}
}
\value{
a data.table containing summary information on target variables specified
}
\description{
summary_table
}
\examples{
x
}
