% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTables.R
\name{summary_table_by}
\alias{summary_table_by}
\title{Create a summary table using multiple rows for grouping on one target column}
\usage{
summary_table_by(
  dt,
  target,
  treat,
  rows_by,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL,
  pct_dec = 2
)
}
\arguments{
\item{dt}{table to perform function on}

\item{target}{vector of column names desired to obtain information on}

\item{treat}{string of treatment variable used for splitting / grouping data}

\item{rows_by}{string, grouping variable to split events by.}

\item{indent}{indent to be used for display and formatting purposes}

\item{.total_dt}{optional table for total counts to be derived}

\item{pct_dec}{decimal places for percentages}
}
\value{
list containing a data.table containing summary information on target variables specified
}
\description{
Create a summary table using multiple rows for grouping on one target column
}
\examples{
adlb <- random.cdisc.data::cadlb|>dplyr::filter(AVISIT != "SCREENING")
labs <- summary_table_by(adlb, 'AVAL', 'ARM', c('PARAM','AVISIT'), '  ', NULL)

}
