% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.R
\name{tern_summary_table}
\alias{tern_summary_table}
\title{Create a clinical reporting table with tern/rtables}
\usage{
tern_summary_table(dt, target, treat, target_name = target)
}
\arguments{
\item{dt}{table to perform function on}

\item{target}{vector of column names desired to obtain information on}

\item{treat}{string of treatment variable used for splitting / grouping data}

\item{target_name}{names to be displayed for target columns either NULL, which uses target, of vector of desired names for all target columns}
}
\value{
a data.table containing summary information on target variables specified
}
\description{
\code{\link[=tern_summary_table]{tern_summary_table()}} is a convenience wrapper around \code{{rtables}} and
\code{{tern}} commands to generate a clinical reporting summary statistics tables
whilst using a similar interface as \code{\link[=summary_table]{summary_table()}}. This can be helpful
for side by side comparisons of the two functions.
}
\examples{
vars <- c('AGE', 'RACE', 'ETHNIC', 'BMRKR1')
var_labels <- c("Age (yr)", "Race", "Ethnicity", "Continous Level Biomarker 1")

# Summary statistics table split by ARM with custom labels.
tern_summary_table(
  adsl,
  target = vars,
  treat = 'ARM',
  target_name = var_labels
)

}
