% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{total_events}
\alias{total_events}
\title{total_events
Identify the total count of treatment groups}
\usage{
total_events(dt, treat, label)
}
\arguments{
\item{dt}{table to perform function on}

\item{treat}{string, treatment population to use as row headers}

\item{label}{row name to be displayed}
}
\value{
list containing data.table of event
}
\description{
total_events
Identify the total count of treatment groups
}
\examples{
adae <- random.cdisc.data::cadae |> dplyr::filter(ANL01FL == "Y")
e <- total_events(adae, 'ARM', 'Total number of AEs')
}
