% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calc_stats}
\alias{calc_stats}
\alias{calc_stats.numeric}
\alias{calc_stats.character}
\alias{calc_stats.factor}
\alias{calc_stats.logical}
\title{calc_stats
wrapper of calc_desc and calc_counts}
\usage{
calc_stats(
  dt,
  target,
  target_name,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)

\method{calc_stats}{numeric}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)

\method{calc_stats}{character}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)

\method{calc_stats}{factor}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)

\method{calc_stats}{logical}(
  dt,
  target,
  target_name = NULL,
  treat,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)
}
\arguments{
\item{dt}{table input to use.}

\item{target}{string, column name to calculate count statistics on}

\item{target_name}{string, how to display column name when new table is created, if null, target is used}

\item{treat}{string, treatment population to use as row headers}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}

\item{.total_dt}{optional table for total counts to be derived, works for character, logical and factor}
}
\value{
a list containing a data.table
}
\description{
calc_stats
wrapper of calc_desc and calc_counts
}
\examples{
adsl <- random.cdisc.data::cadsl
age<-calc_stats(adsl,'AGE',treat='ARM')
}
