% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClass.R
\name{check_table}
\alias{check_table}
\alias{check_table.default}
\alias{check_table.data.table}
\title{check_table}
\usage{
check_table(x)

\method{check_table}{default}(x)

\method{check_table}{data.table}(x)
}
\arguments{
\item{x}{table to check class of}
}
\value{
table converted to data.table
}
\description{
check_table
}
\examples{
check_table(x)
}
