% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlg-package.R
\docType{package}
\name{dtlg-package}
\alias{dtlg}
\alias{dtlg-package}
\title{dtlg: A Performance-Focused Package for Clinical Trial Tables}
\description{
Create high-performance clinical reporting tables (TLGs) from ADaM-like inputs. The package provides a consistent, programmatic API to generate common tables such as demographics, adverse event incidence, and laboratory summaries, using 'data.table' for fast aggregation over large populations. Functions support flexible target-variable selection, stratification by treatment, and customizable summary statistics, and return tidy, machine-readable results ready to render with downstream table/formatting packages in analysis pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://AscentSoftware.github.io/dtlg/}
  \item Report bugs at \url{https://github.com/AscentSoftware/dtlg/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.morgado@acuityanalytics.com}

Authors:
\itemize{
  \item Max Ebenezer-Brown \email{max.ebenezer-brown@acuityanalytics.com}
  \item Max Norman
  \item Xinye Li \email{xinye.li@acuityanalytics.com}
  \item Anja Peebles-Brown \email{anna.peebles-brown@acuityanalytics.com}
  \item Ashley Baldry \email{ashley.baldry@acuityanalytics.com}
}

}
\keyword{internal}
