% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{event_count}
\alias{event_count}
\title{event_count
Functionality to count the number of distinct occurrences of an 'event', defined using .filters,
using patient as the identifier.}
\usage{
event_count(dt, patient, treat, label, .filters = NULL, .total_dt = NULL)
}
\arguments{
\item{dt}{table to perform function on}

\item{patient}{string, column for unique patient identifier}

\item{treat}{string, treatment population to use as row headers}

\item{label}{row name to be displayed}

\item{.filters}{filter variables for event written as single string e.g. "DTHFL == 'Y' & SEX == 'M'}

\item{.total_dt}{optional table for total counts to be derived}
}
\value{
list containing data.table of event
}
\description{
event_count
Functionality to count the number of distinct occurrences of an 'event', defined using .filters,
using patient as the identifier.
}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae |> dplyr::filter(ANL01FL == "Y")
a<-event_count(adsl,patient = "USUBJID",treat = "ARM", label = "Total number of deaths", .filters = "DTHFL == 'Y'", .total_dt = NULL)
b<-event_count(adsl,patient = "USUBJID",treat = "ARM", label = "Total number of patients withdrawn from study due to an AE", .filters = "DCSREAS == 'ADVERSE EVENT'", .total_dt = NULL)
d<-event_count(adae, patient = "USUBJID", treat = "ARM", label = "Total number of patients with at least one AE",.total_dt = adsl)
}
