% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{event_count}
\alias{event_count}
\title{event_count}
\usage{
event_count(dt, patient, treat, label, .filters = NULL, .total_dt = NULL)
}
\arguments{
\item{dt}{table to perform function on}

\item{patient}{string, column for unique patient identifier}

\item{treat}{string, treatment population to use as row headers}

\item{label}{row name to be displayed}

\item{.filters}{filter variables for event written as single string e.g. "DTHFL == 'Y' & SEX == 'M'}

\item{.total_dt}{optional table for total counts to be derived}
}
\value{
list containing data.table of event
}
\description{
event_count
}
\examples{
x
}
