% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{multi_event_true}
\alias{multi_event_true}
\title{multi_event_true}
\usage{
multi_event_true(
  dt,
  event_vars,
  patient,
  treat,
  heading,
  label = NULL,
  .total_dt = NULL,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;"
)
}
\arguments{
\item{dt}{table to perform function on}

\item{event_vars}{events to perform counts on}

\item{patient}{string, column for unique patient identifier}

\item{treat}{string, treatment population to use as row headers}

\item{heading}{string, title row to add to the returned table}

\item{label}{row name to be displayed, if NULL then column name or label is used}

\item{.total_dt}{optional table for total counts to be derived}

\item{indent}{indentation to use for statistic row names, used for formatting outputs for shiny}
}
\value{
list of data.table containing events
}
\description{
multi_event_true
}
\examples{
set.seed(99)
adae <- random.cdisc.data::cadae |> dplyr::mutate(
AEDECOD = with_label(as.character(AEDECOD), "Dictionary-Derived Term"),
AESDTH = with_label(
  sample(c("N", "Y"), size = nrow(adae), replace = TRUE, prob = c(0.99, 0.01)),
  "Results in Death"
),
AEACN = with_label(
  sample(
    c("DOSE NOT CHANGED", "DOSE INCREASED", "DRUG INTERRUPTED", "DRUG WITHDRAWN"),
    size = nrow(adae),
    replace = TRUE, prob = c(0.68, 0.02, 0.25, 0.05)
  ),
  "Action Taken with Study Treatment"
),
FATAL = formatters::with_label(AESDTH == "Y", "AE with fatal outcome"),
SEV = formatters::with_label(AESEV == "SEVERE", "Severe AE (at greatest intensity)"),
SER = formatters::with_label(AESER == "Y", "Serious AE"),
SERWD = formatters::with_label(AESER == "Y" & AEACN == "DRUG WITHDRAWN", "Serious AE leading to withdrawal from treatment"),
SERDSM = formatters::with_label(
  AESER == "Y" & AEACN \%in\% c("DRUG INTERRUPTED", "DOSE INCREASED", "DOSE REDUCED"),
  "Serious AE leading to dose modification/interruption"
),
RELSER = formatters::with_label(AESER == "Y" & AEREL == "Y", "Related Serious AE"),
WD = formatters::with_label(AEACN == "DRUG WITHDRAWN", "AE leading to withdrawal from treatment"),
DSM = formatters::with_label(
  AEACN \%in\% c("DRUG INTERRUPTED", "DOSE INCREASED", "DOSE REDUCED"), "AE leading to dose modification/interruption"
),
REL = formatters::with_label(AEREL == "Y", "Related AE"),
RELWD = formatters::with_label(AEREL == "Y" & AEACN == "DRUG WITHDRAWN", "Related AE leading to withdrawal from treatment"),
RELDSM = formatters::with_label(
 AEREL == "Y" & AEACN \%in\% c("DRUG INTERRUPTED", "DOSE INCREASED", "DOSE REDUCED"),
  "Related AE leading to dose modification/interruption"
)) |> dplyr::filter(ANL01FL == "Y")

aesi_vars <- c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL", "RELWD", "RELDSM", "SEV")
f <- multi_event_true(adae, event_vars = aesi_vars, patient = "USUBJID", treat = "ARM", heading = "Total number of patients with at least one", .total_dt = adsl, indent = "  ")
}
