% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTables.R
\name{summary_table}
\alias{summary_table}
\title{summary_table
wrapper function to create data.table on target variables using calc_stats}
\usage{
summary_table(
  dt,
  target,
  treat,
  target_name = NULL,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)
}
\arguments{
\item{dt}{table to perform function on}

\item{target}{vector of column names desired to obtain information on}

\item{treat}{string of treatment variable used for splitting / grouping data}

\item{target_name}{names to be displayed for target columns either NULL, which uses target, of vector of desired names for all target columns}

\item{indent}{indent to be used for display and formatting purposes}

\item{.total_dt}{optional table for total counts to be derived}
}
\value{
a data.table containing summary information on target variables specified
}
\description{
summary_table
wrapper function to create data.table on target variables using calc_stats
}
\examples{
adsl <- random.cdisc.data::cadsl
vars<-c('AGE','RACE','ETHNIC','BMRKR1')
var_labels <- c("Age (yr)","Sex", "Ethnicity", "Race", "Continous Level Biomarker 1")
DMT01<-summary_table(adsl, target = vars, treat = 'ARM', target_name = var_lables, indent = '  ')
DMT01_pct<-summary_table(adsl, target = vars, treat = 'ARM', indent = '  ', .total_dt = adsl)
}
