% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTables.R
\name{summary_table_by}
\alias{summary_table_by}
\title{Create a summary table using multiple rows for grouping, only works for 2 row groups currently.}
\usage{
summary_table_by(
  dt,
  target,
  treat,
  rows_by,
  indent = "&nbsp;&nbsp;&nbsp;&nbsp;",
  .total_dt = NULL
)
}
\arguments{
\item{dt}{table to perform function on}

\item{target}{vector of column names desired to obtain information on}

\item{treat}{string of treatment variable used for splitting / grouping data}

\item{rows_by}{string, grouping variable to split events by.}

\item{indent}{indent to be used for display and formatting purposes}

\item{.total_dt}{optional table for total counts to be derived}
}
\value{
a data.table containing summary information on target variables specified
}
\description{
Create a summary table using multiple rows for grouping, only works for 2 row groups currently.
}
\examples{
adlb <- random.cdisc.data::cadlb
labs <- summary_table_by(adlb, 'AVAL', 'ARM', c('PARAM','AVISIT'), '  ', NULL)

}
