% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withEvent.R
\name{total_events}
\alias{total_events}
\title{total_events}
\usage{
total_events(dt, treat, label)
}
\arguments{
\item{dt}{table to perform function on}

\item{treat}{string, treatment population to use as row headers}

\item{label}{row name to be displayed}
}
\value{
list containing data.table of event
}
\description{
total_events
}
\examples{
x
}
